refactor(wave-a): hoist snapshot retention config into neutral module
`InMemoryBackend` and `SqliteBackend` were both importing
`resolveMaxRecords` (and the `DEFAULT_SNAPSHOT_MAX_RECORDS` constant)
from `projections/store.js`, which inverted the intended layer
boundary — storage sits below projections in the dispatch-core
diagram, so a storage backend should not depend on projection
adapter code.

Move the retention config into a neutral module at
`storage/snapshot-retention.ts`. Both backends and the projections
wrapper now import from there; `projections/store.ts` re-exports the
symbols so existing external consumers keep resolving. The
WARN-on-prune log emission stays at the projections wrapper boundary
so the observable surface is unchanged.

Addresses CodeRabbit comment 3224637373 on PR #1346.
