+++ servers/exarchos-mcp/src/storage/sqlite-backend.ts
@@ -7,7 +7,7 @@ import type { QueryFilters } from '../event-store/store.js';
 import type { StorageBackend, EventSender, ViewCacheEntry, DrainResult } from './backend.js';
 import { VersionConflictError } from './memory-backend.js';
 import type { SnapshotRecord } from '../projections/snapshot-schema.js';
-import { resolveMaxRecords } from '../projections/store.js';
+import { resolveMaxRecords } from './snapshot-retention.js';
 
 // ─── AtomicAppender wire types (#1259, T06/T07) ─────────────────────────────
 //
